# packages/demo-api/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getData",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/scene-objects": {
      "get": {
        "operationId": "SceneObjectController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "SceneObject"
        ]
      },
      "post": {
        "operationId": "SceneObjectController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSceneObjectDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "SceneObject"
        ]
      }
    },
    "/scene-objects/{id}": {
      "get": {
        "operationId": "SceneObjectController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "SceneObject"
        ]
      },
      "put": {
        "operationId": "SceneObjectController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSceneObjectDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "SceneObject"
        ]
      },
      "delete": {
        "operationId": "SceneObjectController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "SceneObject"
        ]
      }
    }
  },
  "info": {
    "title": "Scene API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "Vector3Dto": {
        "type": "object",
        "properties": {
          "x": {
            "type": "number"
          },
          "y": {
            "type": "number"
          },
          "z": {
            "type": "number"
          }
        },
        "required": [
          "x",
          "y",
          "z"
        ]
      },
      "CreateSceneObjectDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "cube",
              "sphere"
            ],
            "description": "Shape type"
          },
          "position": {
            "$ref": "#/components/schemas/Vector3Dto"
          },
          "scale": {
            "$ref": "#/components/schemas/Vector3Dto"
          },
          "color": {
            "type": "string",
            "example": "#ff0000",
            "description": "Hex color code"
          }
        },
        "required": [
          "type",
          "position",
          "scale",
          "color"
        ]
      },
      "UpdateSceneObjectDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "cube",
              "sphere"
            ]
          },
          "position": {
            "$ref": "#/components/schemas/Vector3Dto"
          },
          "scale": {
            "$ref": "#/components/schemas/Vector3Dto"
          },
          "color": {
            "type": "string",
            "example": "#ff0000"
          }
        }
      }
    }
  }
}